Restructuring files a bit.
